+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 - Add `fixture.Dispose()` to dispose all generated results and all disposable customizations
 - Add that the `UnitAutoData` or `IntegrationAutoData` should automatically dispose after test run
 
+### Fixed
+- Fix issue with `fixture.Freeze()` when using `AutoMockTypeControl` together with `SubClassTransformCustomization`
+
 ## [1.1.7] - 2023-08-16
 
 ### Fixed